+++ lib/src/analyzers/lint_analyzer/lint_analysis_config.dart
@@ -16,6 +16,7 @@ class LintAnalysisConfig {
   final Iterable<Metric> methodsMetrics;
   final Iterable<Glob> metricsExcludes;
   final Map<String, Object> metricsConfig;
+  final String excludesRootFolder;
 
   const LintAnalysisConfig(
     this.globalExcludes,
@@ -25,5 +26,6 @@ class LintAnalysisConfig {
     this.methodsMetrics,
     this.metricsExcludes,
     this.metricsConfig,
+    this.excludesRootFolder,
   );
 }